These are the steps of the algorithm:
Initialize all vertices as unvisited.
Select an arbitrary vertex, set it as the current vertex u. Mark u as visited.
Find out the shortest edge connecting the current vertex u and an unvisited vertex v.
Set v as the current vertex u. Mark v as visited.
If all the vertices in the domain are visited, then terminate. Else, go to step 3.